# swagger.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:3000'
basePath: /
tags:
  - name: Node Printer Controller
    description: Controller to the Node Printer Server
paths:
  '/listPrinters':
    get:
      tags:
        - controller
      summary: List Printers
      operationId: listPrintersUsingGET
      produces:
        - application/json;charset=UTF-8
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/defaultPrinterName':
    get:
      tags:
        - controller
      summary: Get Default Printer Name
      operationId: getDefaultPrinterNameUsingGET
      produces:
        - application/json;charset=UTF-8
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/setDefaultPrinter':
    post:
      tags:
        - controller
      summary: Sets the Default Printer on the system
      operationId: setsDefaultPrinterUsingPOST
      consumes:
        - application/json
      produces:
        - application/json;charset=UTF-8
      parameters:
        - in: body
          name: printerName
          description: Printer Name
          required: true
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/printText':
    post:
      tags:
        - controller
      summary: Prints a String using the Default Printer
      operationId: printsTextUsingPOST
      consumes:
        - application/json
      produces:
        - application/json;charset=UTF-8
      parameters:
        - in: body
          name: text
          description: Text to be printed by the default printer
          required: true
          type: string
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/printTextFromPrinter':
    post:
      tags:
        - controller
      summary: Prints a string from a specific printer
      operationId: printsTextFromPrinterUsingPOST
      consumes:
        - application/json
      produces:
        - application/json;charset=UTF-8
      parameters:
        - in: body
          name: text
          description: Text to be printed
          required: true
          type: string
        - in: body
          name: printerName
          description: Name of the Printer to be used
          required: true
          type: string
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found